+++ mototrack-backend-rest-dotnet/Data/Repositories/Interface/IColaboradorRepository.cs
@@ -5,8 +5,8 @@ namespace mototrack_backend_rest_dotnet.Data.Repositories.Interface;
 public interface IColaboradorRepository
 {
     Task<PageResultModel<IEnumerable<ColaboradorEntity>>> ObterTodosColaboradoresAsync(int deslocamento = 0, int registrosRetornados = 10);
-    Task<ColaboradorEntity?> ObterColaboradorPorIdAsync(int id);
+    Task<ColaboradorEntity?> ObterColaboradorPorIdAsync(long id);
     Task<ColaboradorEntity?> AdicionarColaboradorAsync(ColaboradorEntity colaborador);
-    Task<ColaboradorEntity?> EditarColaboradorAsync(int id, ColaboradorEntity novoColaborador);
-    Task<ColaboradorEntity?> DeletarColaboradorAsync(int id);
+    Task<ColaboradorEntity?> EditarColaboradorAsync(long id, ColaboradorEntity novoColaborador);
+    Task<ColaboradorEntity?> DeletarColaboradorAsync(long id);
 }
