+++ mototrack-backend-rest-dotnet/Services/Interface/IColaboradorService.cs
@@ -6,8 +6,8 @@ namespace mototrack_backend_rest_dotnet.Services.Interface;
 public interface IColaboradorService
 {
     Task<PageResultModel<IEnumerable<ColaboradorEntity>>> ObterTodosColaboradoresAsync(int deslocamento = 0, int registrosRetornados = 10);
-    Task<ColaboradorEntity?> ObterColaboradorPorIdAsync(int id);
+    Task<ColaboradorEntity?> ObterColaboradorPorIdAsync(long id);
     Task<ColaboradorEntity?> AdicionarColaboradorAsync(ColaboradorDTO colaboradorDTO);
-    Task<ColaboradorEntity?> EditarColaboradorAsync(int id, ColaboradorDTO novoColaboradorDTO);
-    Task<ColaboradorEntity?> DeletarColaboradorAsync(int id);
+    Task<ColaboradorEntity?> EditarColaboradorAsync(long id, ColaboradorDTO novoColaboradorDTO);
+    Task<ColaboradorEntity?> DeletarColaboradorAsync(long id);
 }
